Fix crash on March 2024 security patch devices
The March security patch changes the method signatures
of freezeRotation and thawRotation of IWindowManager.
A new String "caller" is expected.

When looking up freezeRotation if the first attempt
fails try again with a String.class added to the argument list
When looking up thawRotation if the first attempt
fails try again with a String.class added to the argument list
When invoking these methods, first check if newMethods is true
if so pass the apps FQN as the caller argument.

Signed-off-by: Brock York <twunknown@gmail.com>